@@ -21,44 +21,54 @@
     "build": "ember build",
     "build-dev": "ember build --environment=production",
     "build-prod": "ember build --environment=production",
+    "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*:fix",
     "lint:hbs": "ember-template-lint .",
-    "lint:js": "eslint .",
+    "lint:hbs:fix": "ember-template-lint . --fix",
+    "lint:js": "eslint . --cache",
+    "lint:js:fix": "eslint . --fix",
     "start": "ember serve",
-    "test": "ember test"
+    "test": "npm-run-all lint test:*"
   },
   "devDependencies": {
-    "@ember/optional-features": "^1.1.0",
-    "@ember/render-modifiers": "^1.0.2",
-    "@embroider/compat": "^0.28.0",
-    "@embroider/core": "^0.28.0",
-    "@embroider/webpack": "^0.28.0",
-    "@glimmer/component": "^1.0.2",
+    "@ember/optional-features": "^2.0.0",
+    "@ember/test-helpers": "^2.6.0",
+    "@embroider/compat": "^1.1.0",
+    "@embroider/core": "^1.1.0",
+    "@embroider/webpack": "^1.1.0",
+    "@glimmer/component": "^1.0.4",
     "babel-eslint": "^10.1.0",
     "broccoli-asset-rev": "3.0.0",
-    "ember-auto-import": "^1.6.0",
-    "ember-cli": "~3.22.0",
-    "ember-cli-babel": "^7.23.0",
+    "ember-auto-import": "^2.4.0",
+    "ember-cli": "~4.1.1",
+    "ember-cli-babel": "^7.26.11",
     "ember-cli-dependency-checker": "3.2.0",
-    "ember-cli-eslint": "5.1.0",
-    "ember-cli-htmlbars": "^4.2.0",
-    "ember-cli-inject-live-reload": "2.0.2",
+    "ember-cli-htmlbars": "^6.0.1",
+    "ember-cli-inject-live-reload": "2.1.0",
     "ember-cli-sri": "2.1.1",
-    "ember-cli-template-trimmer": "0.0.6",
-    "ember-cli-uglify": "3.0.0",
-    "ember-cli-update": "0.54.6",
-    "ember-hbs-minifier": "0.5.0",
-    "ember-load-initializers": "^2.1.1",
-    "ember-qunit": "^4.6.0",
-    "ember-resolver": "^7.0.0",
-    "ember-source": "3.22.0",
-    "eslint-plugin-ember": "^7.7.1",
+    "ember-cli-terser": "4.0.2",
+    "ember-cli-update": "^1.0.1",
+    "ember-load-initializers": "^2.1.2",
+    "ember-qunit": "^5.1.5",
+    "ember-resolver": "^8.0.3",
+    "ember-source": "4.2.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-ember": "^10.5.8",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-qunit": "^7.2.0",
     "loader.js": "4.7.0",
-    "qunit-dom": "^0.9.2"
+    "prettier": "^2.5.1",
+    "qunit": "^2.17.2",
+    "qunit-dom": "^2.0.0",
+    "webpack": "^5.68.0"
   },
   "engines": {
-    "node": "6.* || 8.* || >= 10.*"
+    "node": "12.* || 14.* || >= 16"
   },
   "ember": {
     "edition": "octane"
+  },
+  "dependencies": {
+    "ember-functions-as-modifiers-polyfill": "^1.0.0"
   }
 }
