@@ -1,6 +1,6 @@
 {
   "name": "js-framework-benchmark-ls-element",
-  "version": "1.1.7",
+  "version": "2.0.9",
   "description": "ls-element demo",
   "main": "index.js",
   "js-framework-benchmark": {
@@ -31,12 +31,12 @@
     "html-webpack-plugin": "^4.3.0",
     "terser-webpack-plugin": "^4.1.0",
     "ts-loader": "^8.0.2",
-    "typescript": "^3.9.6",
+    "typescript": "^4.4.4",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
-    "@lsegurado/ls-element": "1.1.7"
+    "@lsegurado/ls-element": "2.0.9"
   }
 }
