@@ -16,9 +16,9 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@wasm-tool/wasm-pack-plugin": "^0.4.2",
-    "webpack": "^4.33.0",
-    "webpack-cli": "^3.3.3",
-    "rimraf": "^2.6.3"
+    "@wasm-tool/wasm-pack-plugin": "^1.5.0",
+    "webpack": "^5.51.1",
+    "webpack-cli": "^4.8.0",
+    "rimraf": "^3.0.2"
   }
 }
