@@ -51,11 +51,20 @@ Doo.define(
         }
 
 		buildData(count = 1000) {
-			const data = [];
+			const data = []
 			for (let i = 0; i < count; i++) {
 				data.push({id: this.ID++,label: adjectives[_random(lenA)] + " " + colours[_random(lenB)] + " " + nouns[_random(lenC)]})
 			}
 			return data	
+		}
+	
+		getIndex(row) {
+			let idx =  this.data.rows.findIndex((item, i) => {
+				if (item.id === row.key) {
+					return i
+				}
+			}) 
+			return idx
 		}
 
 		getIndex(row) {
@@ -70,6 +79,7 @@ Doo.define(
 		delete(elem) {
 			let row = this.getParentRow(elem)
 			if (row) {
+				let idx = this.getIndex(row)
 				this.tbody.removeChild(row)
 				let idx = this.getIndex(row)
 				if (idx !== undefined) {
@@ -111,6 +121,7 @@ Doo.define(
 		}
 
 		toggleSelect(row) {
+			let row = this.getParentRow(elem)
 			if (row) {
 				row.classList.toggle('danger')
 				if (row.classList.contains('danger')) {
@@ -133,6 +144,7 @@ Doo.define(
 					row1 = this.data.rows[1]
 				
 				this.data.rows[1] = this.data.rows[998];
+
 				this.data.rows[998] = row1
 				
 				this.tbody.insertBefore(node998, node2)
