@@ -1,18 +1,11 @@
 'use strict';
 
-module.exports = function(environment) {
+module.exports = function (environment) {
   let ENV = {
-
-    'ember-resolver': {
-      features: {
-        // EMBER_RESOLVER_MODULE_UNIFICATION: true,
-        // EMBER_GLIMMER_ANGLE_BRACKET_INVOCATION: true,
-      },
-    },
     modulePrefix: 'ember-temp',
     environment,
     rootURL: '/frameworks/keyed/ember/dist/',
-    locationType: 'auto',
+    locationType: 'history',
     EmberENV: {
       FEATURES: {
         // Here you can enable experimental features on an ember canary build
@@ -23,14 +16,14 @@ module.exports = function(environment) {
       },
       EXTEND_PROTOTYPES: {
         // Prevent Ember Data from overriding Date.parse.
-        Date: false
-      }
+        Date: false,
+      },
     },
 
     APP: {
       // Here you can pass flags/options to your application instance
       // when it is created
-    }
+    },
   };
 
   if (environment === 'development') {
