@@ -3,9 +3,9 @@
 const browsers = [
   'last 1 Chrome versions',
   'last 1 Firefox versions',
-  'last 1 Safari versions'
+  'last 1 Safari versions',
 ];
 
 module.exports = {
-  browsers
+  browsers,
 };
