@@ -1,8 +1,8 @@
-import { createTemplate, setComponentTemplate, templateOnlyComponent } from '@glimmer/core';
+import { precompileTemplate, setComponentTemplate, templateOnlyComponent } from '@glimmer/core';
 
 import MyTable from './MyTable';
 
 export default setComponentTemplate(
-  createTemplate({ MyTable }, `<div class="container"><MyTable /></div>`),
+  precompileTemplate(`<div class="container"><MyTable /></div>`,  { scope: () => ({ MyTable }), strictMode: true }),
   templateOnlyComponent()
 );
