@@ -1,6 +1,6 @@
-import { createTemplate, setComponentTemplate, templateOnlyComponent } from '@glimmer/core';
+import { precompileTemplate, setComponentTemplate, templateOnlyComponent } from '@glimmer/core';
 
 export default setComponentTemplate(
-  createTemplate(`<button type="button" class="btn btn-primary btn-block" ...attributes>{{yield}}</button>`),
+  precompileTemplate(`<button type="button" class="btn btn-primary btn-block" ...attributes>{{yield}}</button>`,  { strictMode: true }),
   templateOnlyComponent()
 );
