@@ -24,7 +24,7 @@ module.exports = {
               ['@babel/plugin-proposal-decorators', { legacy: true }],
               '@babel/plugin-proposal-class-properties',
             ],
-            presets: ['@babel/preset-typescript', '@babel/preset-env'],
+            presets: ['@babel/preset-typescript', '@babel/preset-env', '@glimmer/babel-preset'],
           },
         },
       },
