@@ -1,5 +1,5 @@
 import { createSignal, createSelector, batch } from 'solid-js';
-import { render } from 'solid-js/dom';
+import { render } from 'solid-js/web';
 
 let idCounter = 1;
 const adjectives = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"],
@@ -27,7 +27,28 @@ const Button = ({ id, text, fn }) =>
 
 const App = () => {
   const [data, setData] = createSignal([]),
-    [selected, setSelected] = createSignal(null, true),
+    [selected, setSelected] = createSignal(null),
+    run = () => setData(buildData(1000)),
+    runLots = () => setData(buildData(10000)),
+    add = () => setData(d => [...d, ...buildData(1000)]),
+    update = () => batch(() => {
+      for(let i = 0, d = data(), len = d.length; i < len; i += 10)
+        d[i].setLabel(l => l + ' !!!');
+    }),
+    swapRows = () => {
+      const d = data().slice();
+      if (d.length > 998) {
+        let tmp = d[1];
+        d[1] = d[998];
+        d[998] = tmp;
+        setData(d);
+      }
+    },
+    clear = () => setData([]),
+    remove = id => setData(d => {
+      const idx = d.findIndex(d => d.id === id);
+      return [...d.slice(0, idx), ...d.slice(idx + 1)];
+    }),
     isSelected = createSelector(selected);
 
   return <div class='container'>
@@ -55,56 +76,6 @@ const App = () => {
     </tbody></table>
     <span class='preloadicon glyphicon glyphicon-remove' aria-hidden="true" />
   </div>;
-
-  function remove(id) {
-    const d = data();
-    d.splice(d.findIndex(d => d.id === id), 1);
-    setData(d);
-  }
-
-  function run() {
-    batch(() => {
-      setData(buildData(1000));
-      setSelected(null);
-    });
-  }
-
-  function runLots() {
-    batch(() => {
-      setData(buildData(10000));
-      setSelected(null);
-    });
-  }
-
-  function add() { setData(data().concat(buildData(1000))); }
-
-  function update() {
-    batch(() => {
-      const d = data();
-      let index = 0;
-      while (index < d.length) {
-        d[index].setLabel(d[index].label() + ' !!!');
-        index += 10;
-      }
-    });
-  }
-
-  function swapRows() {
-    const d = data();
-    if (d.length > 998) {
-      let tmp = d[1];
-      d[1] = d[998];
-      d[998] = tmp;
-      setData(d);
-    }
-  }
-
-  function clear() {
-    batch(() => {
-      setData([]);
-      setSelected(null);
-    });
-  }
 }
 
 render(App, document.getElementById("main"));
