@@ -1,7 +1,7 @@
 import minifyHTML from 'rollup-plugin-minify-html-literals';
-import resolve from 'rollup-plugin-node-resolve';
+import {nodeResolve} from '@rollup/plugin-node-resolve';
+import {terser} from 'rollup-plugin-terser';
 import includePaths from 'rollup-plugin-includepaths';
-import { terser } from 'rollup-plugin-terser';
 
 export default {
   input: 'src/index.js',
@@ -15,17 +15,19 @@ export default {
     }),
     includePaths({
       include: {
-        '@ungap/create-content': './node_modules/@ungap/degap/create-content.js'
+        '@ungap/create-content': 'node_modules/@ungap/degap/create-content.js'
       },
     }),
-    resolve(),
+    nodeResolve(),
     terser()
   ],
-  context: 'null',
-  moduleContext: 'null',
   output: {
+    esModule: false,
     file: 'dist/index.js',
+    exports: 'named',
     format: 'iife',
     name: 'app'
   }
 };
+
+
