@@ -4,16 +4,12 @@ import {html, render} from 'uhtml';
 import Jumbotron from './jumbotron.js';
 import Table from './table.js';
 
-const state = State(update);
-const main = document.getElementById('container');
+const state = State(Table, false, html.for);
 
-update(state);
-
-function update(state) {
-  render(main, html`
+render(document.getElementById('container'), html`
   <div class="container">
     ${Jumbotron(state)}
     ${Table(state)}
     <span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true" />
-  </div>`);
-}
+  </div>
+`);
