@@ -1,1 +1,1 @@
-!function(e){function _(_){for(var n,r,o=_[0],i=_[1],u=0,f=[];u<o.length;u++)r=o[u],t[r]&&f.push(t[r][0]),t[r]=0;for(n in i)Object.prototype.hasOwnProperty.call(i,n)&&(e[n]=i[n]);for(d&&d(_);f.length;)f.shift()()}var n={},t={0:0};var r={};var o={2:function(){return{"./index.js":{__widl_f_tag_name_Element:function(e,_){return n[1].exports.__widl_f_tag_name_Element(e,_)},__widl_f_parent_element_Node:function(e){return n[1].exports.__widl_f_parent_element_Node(e)},__wbindgen_object_drop_ref:function(e){return n[1].exports.__wbindgen_object_drop_ref(e)},__widl_f_get_attribute_Element:function(e,_,t,r){return n[1].exports.__widl_f_get_attribute_Element(e,_,t,r)},__widl_f_set_text_content_Node:function(e,_,t){return n[1].exports.__widl_f_set_text_content_Node(e,_,t)},__widl_f_set_class_name_Element:function(e,_,t){return n[1].exports.__widl_f_set_class_name_Element(e,_,t)},__wbg_random_4eec15c1337c828c:function(){return n[1].exports.__wbg_random_4eec15c1337c828c()},__widl_f_clone_node_with_deep_Node:function(e,_){return n[1].exports.__widl_f_clone_node_with_deep_Node(e,_)},__widl_f_first_child_Node:function(e){return n[1].exports.__widl_f_first_child_Node(e)},__widl_f_next_sibling_Node:function(e){return n[1].exports.__widl_f_next_sibling_Node(e)},__widl_f_set_attribute_Element:function(e,_,t,r,o){return n[1].exports.__widl_f_set_attribute_Element(e,_,t,r,o)},__widl_f_append_child_Node:function(e,_){return n[1].exports.__widl_f_append_child_Node(e,_)},__widl_f_target_Event:function(e){return n[1].exports.__widl_f_target_Event(e)},__widl_f_id_Element:function(e,_){return n[1].exports.__widl_f_id_Element(e,_)},__widl_f_prevent_default_Event:function(e){return n[1].exports.__widl_f_prevent_default_Event(e)},__widl_f_class_list_Element:function(e){return n[1].exports.__widl_f_class_list_Element(e)},__widl_f_contains_DOMTokenList:function(e,_,t){return n[1].exports.__widl_f_contains_DOMTokenList(e,_,t)},__wbindgen_object_clone_ref:function(e){return n[1].exports.__wbindgen_object_clone_ref(e)},__widl_f_insert_before_Node:function(e,_,t){return n[1].exports.__widl_f_insert_before_Node(e,_,t)},__widl_f_remove_Element:function(e){return n[1].exports.__widl_f_remove_Element(e)},__wbg_newnoargs_8d1797b163dbc9fb:function(e,_){return n[1].exports.__wbg_newnoargs_8d1797b163dbc9fb(e,_)},__wbg_call_836fa928f74337e5:function(e,_){return n[1].exports.__wbg_call_836fa928f74337e5(e,_)},__widl_instanceof_Window:function(e){return n[1].exports.__widl_instanceof_Window(e)},__widl_f_document_Window:function(e){return n[1].exports.__widl_f_document_Window(e)},__widl_f_create_element_Document:function(e,_,t){return n[1].exports.__widl_f_create_element_Document(e,_,t)},__widl_f_set_inner_html_Element:function(e,_,t){return n[1].exports.__widl_f_set_inner_html_Element(e,_,t)},__widl_f_get_element_by_id_Document:function(e,_,t){return n[1].exports.__widl_f_get_element_by_id_Document(e,_,t)},__widl_f_add_event_listener_with_callback_EventTarget:function(e,_,t,r){return n[1].exports.__widl_f_add_event_listener_with_callback_EventTarget(e,_,t,r)},__wbindgen_cb_forget:function(e){return n[1].exports.__wbindgen_cb_forget(e)},__wbindgen_cb_drop:function(e){return n[1].exports.__wbindgen_cb_drop(e)},__wbindgen_debug_string:function(e,_){return n[1].exports.__wbindgen_debug_string(e,_)},__wbindgen_throw:function(e,_){return n[1].exports.__wbindgen_throw(e,_)},__wbindgen_rethrow:function(e){return n[1].exports.__wbindgen_rethrow(e)},__wbindgen_closure_wrapper60:function(e,_,t){return n[1].exports.__wbindgen_closure_wrapper60(e,_,t)}}}}};function i(_){if(n[_])return n[_].exports;var t=n[_]={i:_,l:!1,exports:{}};return e[_].call(t.exports,t,t.exports,i),t.l=!0,t.exports}i.e=function(e){var _=[],n=t[e];if(0!==n)if(n)_.push(n[2]);else{var u=new Promise(function(_,r){n=t[e]=[_,r]});_.push(n[2]=u);var f,l=document.createElement("script");l.charset="utf-8",l.timeout=120,i.nc&&l.setAttribute("nonce",i.nc),l.src=function(e){return i.p+""+({}[e]||e)+".js"}(e);var d=new Error;f=function(_){l.onerror=l.onload=null,clearTimeout(c);var n=t[e];if(0!==n){if(n){var r=_&&("load"===_.type?"missing":_.type),o=_&&_.target&&_.target.src;d.message="Loading chunk "+e+" failed.\n("+r+": "+o+")",d.name="ChunkLoadError",d.type=r,d.request=o,n[1](d)}t[e]=void 0}};var c=setTimeout(function(){f({type:"timeout",target:l})},12e4);l.onerror=l.onload=f,document.head.appendChild(l)}return({1:[2]}[e]||[]).forEach(function(e){var n=r[e];if(n)_.push(n);else{var t,u=o[e](),f=fetch(i.p+""+{2:"9514be0dfb95a4ff4996"}[e]+".module.wasm");if(u instanceof Promise&&"function"==typeof WebAssembly.compileStreaming)t=Promise.all([WebAssembly.compileStreaming(f),u]).then(function(e){return WebAssembly.instantiate(e[0],e[1])});else if("function"==typeof WebAssembly.instantiateStreaming)t=WebAssembly.instantiateStreaming(f,u);else{t=f.then(function(e){return e.arrayBuffer()}).then(function(e){return WebAssembly.instantiate(e,u)})}_.push(r[e]=t.then(function(_){return i.w[e]=(_.instance||_).exports}))}}),Promise.all(_)},i.m=e,i.c=n,i.d=function(e,_,n){i.o(e,_)||Object.defineProperty(e,_,{enumerable:!0,get:n})},i.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},i.t=function(e,_){if(1&_&&(e=i(e)),8&_)return e;if(4&_&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(i.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&_&&"string"!=typeof e)for(var t in e)i.d(n,t,function(_){return e[_]}.bind(null,t));return n},i.n=function(e){var _=e&&e.__esModule?function(){return e.default}:function(){return e};return i.d(_,"a",_),_},i.o=function(e,_){return Object.prototype.hasOwnProperty.call(e,_)},i.p="bundled-dist/",i.oe=function(e){throw console.error(e),e},i.w={};var u=window.webpackJsonp=window.webpackJsonp||[],f=u.push.bind(u);u.push=_,u=u.slice();for(var l=0;l<u.length;l++)_(u[l]);var d=f;i(i.s=0)}([function(e,_,n){n.e(1).then(n.bind(null,1)).catch(console.error)}]);
+(()=>{var e,r,t,n,o,a,i,s={},l={};function c(e){var r=l[e];if(void 0!==r)return r.exports;var t=l[e]={id:e,loaded:!1,exports:{}};return s[e](t,t.exports,c),t.loaded=!0,t.exports}c.m=s,e="function"==typeof Symbol?Symbol("webpack then"):"__webpack_then__",r="function"==typeof Symbol?Symbol("webpack exports"):"__webpack_exports__",t=e=>{e&&(e.forEach((e=>e.r--)),e.forEach((e=>e.r--?e.r++:e())))},n=e=>!--e.r&&e(),o=(e,r)=>e?e.push(r):n(r),c.a=(a,i,s)=>{var l,c,u,d=s&&[],p=a.exports,b=!0,f=!1,m=(r,t,n)=>{f||(f=!0,t.r+=r.length,r.map(((r,o)=>r[e](t,n))),f=!1)},h=new Promise(((e,r)=>{u=r,c=()=>(e(p),t(d),d=0)}));h[r]=p,h[e]=(e,r)=>{if(b)return n(e);l&&m(l,e,r),o(d,e),h.catch(r)},a.exports=h,i((a=>{if(!a)return c();var i,s;l=(a=>a.map((a=>{if(null!==a&&"object"==typeof a){if(a[e])return a;if(a.then){var i=[];a.then((e=>{s[r]=e,t(i),i=0}));var s={};return s[e]=(e,r)=>(o(i,e),a.catch(r)),s}}var l={};return l[e]=e=>n(e),l[r]=a,l})))(a);var u=new Promise(((e,t)=>{(i=()=>e(s=l.map((e=>e[r])))).r=0,m(l,i,t)}));return i.r?u:s})).then(c,u),b=!1},c.d=(e,r)=>{for(var t in r)c.o(r,t)&&!c.o(e,t)&&Object.defineProperty(e,t,{enumerable:!0,get:r[t]})},c.f={},c.e=e=>Promise.all(Object.keys(c.f).reduce(((r,t)=>(c.f[t](e,r),r)),[])),c.u=e=>e+".js",c.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),c.hmd=e=>((e=Object.create(e)).children||(e.children=[]),Object.defineProperty(e,"exports",{enumerable:!0,set:()=>{throw new Error("ES Modules may not assign module.exports or exports.*, Use ESM export syntax, instead: "+e.id)}}),e),c.o=(e,r)=>Object.prototype.hasOwnProperty.call(e,r),a={},i="js-framework-benchmark-wasm-bindgen:",c.l=(e,r,t,n)=>{if(a[e])a[e].push(r);else{var o,s;if(void 0!==t)for(var l=document.getElementsByTagName("script"),u=0;u<l.length;u++){var d=l[u];if(d.getAttribute("src")==e||d.getAttribute("data-webpack")==i+t){o=d;break}}o||(s=!0,(o=document.createElement("script")).charset="utf-8",o.timeout=120,c.nc&&o.setAttribute("nonce",c.nc),o.setAttribute("data-webpack",i+t),o.src=e),a[e]=[r];var p=(r,t)=>{o.onerror=o.onload=null,clearTimeout(b);var n=a[e];if(delete a[e],o.parentNode&&o.parentNode.removeChild(o),n&&n.forEach((e=>e(t))),r)return r(t)},b=setTimeout(p.bind(null,void 0,{type:"timeout",target:o}),12e4);o.onerror=p.bind(null,o.onerror),o.onload=p.bind(null,o.onload),s&&document.head.appendChild(o)}},c.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},c.v=(e,r,t,n)=>{var o=fetch(c.p+""+t+".module.wasm");return"function"==typeof WebAssembly.instantiateStreaming?WebAssembly.instantiateStreaming(o,n).then((r=>Object.assign(e,r.instance.exports))):o.then((e=>e.arrayBuffer())).then((e=>WebAssembly.instantiate(e,n))).then((r=>Object.assign(e,r.instance.exports)))},c.p="bundled-dist/",(()=>{var e={826:0};c.f.j=(r,t)=>{var n=c.o(e,r)?e[r]:void 0;if(0!==n)if(n)t.push(n[2]);else{var o=new Promise(((t,o)=>n=e[r]=[t,o]));t.push(n[2]=o);var a=c.p+c.u(r),i=new Error;c.l(a,(t=>{if(c.o(e,r)&&(0!==(n=e[r])&&(e[r]=void 0),n)){var o=t&&("load"===t.type?"missing":t.type),a=t&&t.target&&t.target.src;i.message="Loading chunk "+r+" failed.\n("+o+": "+a+")",i.name="ChunkLoadError",i.type=o,i.request=a,n[1](i)}}),"chunk-"+r,r)}};var r=(r,t)=>{var n,o,[a,i,s]=t,l=0;if(a.some((r=>0!==e[r]))){for(n in i)c.o(i,n)&&(c.m[n]=i[n]);s&&s(c)}for(r&&r(t);l<a.length;l++)o=a[l],c.o(e,o)&&e[o]&&e[o][0](),e[a[l]]=0},t=self.webpackChunkjs_framework_benchmark_wasm_bindgen=self.webpackChunkjs_framework_benchmark_wasm_bindgen||[];t.forEach(r.bind(null,0)),t.push=r.bind(null,t.push.bind(t))})(),c.e(235).then(c.bind(c,235)).catch(console.error)})();
