@@ -16,8 +16,10 @@ module.exports = {
   },
   plugins: [
     new WasmPackPlugin({
-      crateDirectory: __dirname,
-      extraArgs: "--out-name index"
+      crateDirectory: __dirname
     })
-  ]
+  ],
+  experiments: {
+    asyncWebAssembly: true
+  }
 };
