@@ -6,10 +6,11 @@ interface Props<T> {
   selectAll: (event: React.SyntheticEvent) => void;
   isNoneSelected: boolean;
   areAllSelected: boolean;
+  grid?: boolean;
 }
 
 export function DropDownContents<T>(props: Props<T>) {
-  const {selectNone, selectAll, isNoneSelected, areAllSelected, children} = props;
+  const {selectNone, selectAll, isNoneSelected, areAllSelected, children, grid = false} = props;
   return <div className="section">
             {children[0]}
             <div className="float-rt">
@@ -17,7 +18,7 @@ export function DropDownContents<T>(props: Props<T>) {
                 &nbsp;
                 {!areAllSelected ? <button className='textButton' onClick={selectAll}>All</button> : <button disabled={true} className='textButton'>All</button>}
             </div>
-            <div className="grid">
+            <div className={grid ? "grid" : ""}>
                 {children[1]}
             </div>
         </div>;
