@@ -1,21 +1,26 @@
-import * as fs from 'fs';
-import {JSONResult, config, initializeFrameworks} from './common'
-import {BenchmarkType, Benchmark, benchmarks} from './benchmarks'
+import * as fs from "fs";
+import { JSONResult, config, initializeFrameworks } from "./common";
 
 async function main() {
-	let frameworks = await initializeFrameworks();
-	
-	frameworks.sort( (a,b) => a.fullNameWithKeyedAndVersion.localeCompare(b.fullNameWithKeyedAndVersion));
-	
-	const dots = require('dot').process({
-		path: './'
-	});
-	
-	fs.writeFileSync('../index.html', dots.index({
-		frameworks
-	}), {
-		encoding: 'utf8'
-	})
+  let frameworks = await initializeFrameworks();
+
+  frameworks.sort((a, b) =>
+    a.fullNameWithKeyedAndVersion.localeCompare(b.fullNameWithKeyedAndVersion)
+  );
+
+  const dots = require("dot").process({
+    path: "./",
+  });
+
+  fs.writeFileSync(
+    "../index.html",
+    dots.index({
+      frameworks,
+    }),
+    {
+      encoding: "utf8",
+    }
+  );
 }
 
 main();
